requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.1.0
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1.7
django-fontawesome-5==1.0.18
django-gunicorn==0.1.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.9.0
frozenlist==1.3.3
gunicorn==20.1.0
httpie==3.2.2
idna==3.4
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
openai==0.27.2
openpyxl==3.1.1
Pillow==9.5.0
pipenv==2023.2.18
platformdirs==3.1.1
psycopg2==2.9.6
pycodestyle==2.10.0
Pygments==2.16.1
PySocks==1.7.1
pytz==2023.3
redis==4.5.4
requests==2.28.2
requests-toolbelt==1.0.0
rich==13.5.2
sqlparse==0.4.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
yarl==1.8.2
